# registration_system_go/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Registration System Go API
  version: 1.0.0
  description: 赛事报名与球队管理 Go 后端 HTTP API。
servers:
  - url: http://127.0.0.1:18081
    description: 本地开发服务
...
paths:
...
  /api/v1/app/teams/{id}:
    get:
      operationId: getAppTeam
      tags: [App Teams]
      summary: 获取球队详情
      security: *bearerSecurity
      parameters:
        - $ref: "#/components/parameters/TeamID"
      responses:
        "200":
          description: 球队详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppTeamDetailEnvelope"
        <<: *standardErrors
    patch:
      operationId: updateAppTeamProfile
      tags: [App Teams]
      summary: 更新球队资料（队长/领队）
      description: 仅该队 active 的队长或领队可操作。name 不传则保持不变，传入时必须非空且不超过 120 字符；description/logo_url 传空串或 null 视为清除。球队 status（冻结/激活）不可通过此接口修改。
      security: *bearerSecurity
      parameters:
        - $ref: "#/components/parameters/TeamID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, maxLength: 120 }
                description: { type: string, nullable: true }
                logo_url: { type: string, nullable: true }
      responses:
        "200":
          description: 更新成功（空 data）
        <<: *standardErrors
    delete:
      operationId: deleteAppTeam
      tags: [App Teams]
      summary: 解散球队（仅队长）
      description: 仅该队 active 的队长本人可操作。软删除——teams.status 置为 dissolved，保留历史比赛/申请/支付数据。球队仍有未结束比赛（主/客队）或进行中约队申请时返回 409，可先调 GET /api/v1/app/teams/{id}/dissolve-blockers 查看阻塞项。
      security: *bearerSecurity
      parameters:
        - $ref: "#/components/parameters/TeamID"
      responses:
        "200":
          description: 解散成功（空 data）
        <<: *standardErrors
...
  /api/v1/app/teams/{id}/members:
    get:
      operationId: listAppTeamMembers
      tags: [App Teams]
      summary: 获取球队成员
      security: *bearerSecurity
      parameters:
        - $ref: "#/components/parameters/TeamID"
      responses:
        "200":
          description: 球队成员列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppTeamMemberListEnvelope"
        <<: *standardErrors
    post:
      operationId: addAppTeamMember
      tags: [App Teams]
      summary: 添加球队成员（队长/领队）
      description: 仅该队 active 的队长或领队可操作。role 仅支持 leader/vice_captain/member（缺省为 member），不能直接指定队长；目标用户必须存在且为 active。
      security: *bearerSecurity
      parameters:
        - $ref: "#/components/parameters/TeamID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id: { type: integer, format: int64 }
                role: { type: string, enum: [leader, vice_captain, member] }
      responses:
        "200":
          description: 添加成功（空 data）
        <<: *standardErrors
...
  /api/v1/app/teams/{id}/members/{user_id}/attendance:
    get:
      operationId: getAppTeamMemberAttendance
      tags: [App Teams]
      summary: 获取队员比赛出勤明细（队长/领队）
      security: *bearerSecurity
      parameters:
        - $ref: "#/components/parameters/TeamID"
        - name: user_id
          in: path
          required: true
          schema: { type: integer, format: int64 }
        - name: startDate
          in: query
          required: false
          schema: { type: string, format: date }
        - name: endDate
          in: query
          required: false
          schema: { type: string, format: date }
      responses:
        "200":
          description: 队员在全部已结束比赛中的出勤明细
        <<: *standardErrors
  /api/v1/app/teams/{id}/attendance-summary:
    get:
      operationId: getAppTeamAttendanceSummary
      tags: [App Teams]
      summary: 获取球队出勤摘要（我的记录 + 全队排名，队员可看）
      security: *bearerSecurity
      parameters:
        - $ref: "#/components/parameters/TeamID"
        - name: startDate
          in: query
          required: false
          schema: { type: string, format: date }
        - name: endDate
          in: query
          required: false
          schema: { type: string, format: date }
      responses:
        "200":
          description: 当前队员的出勤记录与全队出勤排名
        <<: *standardErrors
  /api/v1/app/teams/{id}/matches/{match_id}/attendance:
    get:
      operationId: getAppTeamMatchAttendance
      tags: [App Teams]
      summary: 获取单场比赛的全队出勤（队长/领队，仅在职成员）
      security: *bearerSecurity
      parameters:
        - $ref: "#/components/parameters/TeamID"
        - name: match_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: 该场比赛全体在职成员的报名与出勤状态
        <<: *standardErrors
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TeamID:
      name: id
      in: path
      required: true
      description: 球队 ID
      schema:
        type: integer
        format: int64
        minimum: 1
...
  schemas:
    SuccessBase:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          format: int32
          enum: [0]
        message:
          type: string
          example: ok
...
    TeamStatus:
      type: string
      enum: [active, frozen, dissolved]
    MemberStatus:
      type: string
      enum: [active, inactive]
    MemberRole:
      type: string
      enum: [captain, leader, vice_captain, member]
...
    AppTeamDetail:
      type: object
      required: [id, name, description, logo_url, captain_id, status, my_role]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
          nullable: true
        logo_url:
          type: string
          nullable: true
        captain_id:
          type: integer
          format: int64
          nullable: true
        status:
          $ref: "#/components/schemas/TeamStatus"
        my_role:
          $ref: "#/components/schemas/MemberRole"
    AppTeamMember:
      type: object
      required: [user_id, nickname, avatar_url, real_name, role, status, joined_at, balance_cents, is_paid_member, last_recharge_at]
      properties:
        user_id:
          type: integer
          format: int64
        nickname:
          type: string
        avatar_url:
          type: string
          nullable: true
        real_name:
          type: string
          nullable: true
        role:
          $ref: "#/components/schemas/MemberRole"
        status:
          $ref: "#/components/schemas/MemberStatus"
        joined_at:
          type: string
          format: date-time
        balance_cents:
          type: integer
          format: int64
        is_paid_member:
          type: boolean
        last_recharge_at:
          type: string
          format: date-time
          nullable: true
...
    AppTeamDetailEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessBase"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/AppTeamDetail"
...
    AppTeamMemberListEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessBase"
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/AppTeamMember"
...