# registration_system_mini/src/types/app.ts
export type AppUserStatus = "active" | "frozen";

export interface AppUser {
  id: number;
  nickname: string;
  avatar_url: string | null;
  real_name: string | null;
  phone_number: string | null;
  status: AppUserStatus;
  /** 比赛管理员（管理端设置）：可在比赛详情页录入比赛比分。 */
  is_match_admin?: boolean;
}

export interface LoginResponse {
  token: string;
  user: AppUser;
}

export interface ImpersonationTargetsResponse {
  items: AppUser[];
}

export interface WebViewCodeResponse {
  code: string;
  expires_at: string;
}

export interface WebViewExchangeResponse {
  token: string;
}

export type MyTeamRole = "captain" | "leader" | "vice_captain" | "member";

export interface MyTeam {
  id: number;
  name: string;
  description: string | null;
  logo_url: string | null;
  role: MyTeamRole;
  joined_at: string;
  /** 活跃成员数；/teams/my 列表接口直接返回，避免未加载详情的球队显示 0 人。 */
  member_count?: number;
}

export interface AppTeamDetail {
  id: number;
  name: string;
  description: string | null;
  logo_url: string | null;
  captain_id: number | null;
  status: string;
  my_role: MyTeamRole;
}

export interface AppTeamMember {
  user_id: number;
  nickname: string;
  avatar_url: string | null;
  real_name: string | null;
  role: MyTeamRole;
  status: string;
  joined_at: string;
  balance_cents: number;
  is_paid_member: boolean;
  last_recharge_at: string | null;
}

export interface TestLoginUser {
  id: number;
  display_name: string;
  avatar_url: string | null;
  teams: Array<{ id: number; name: string; role: MyTeamRole }>;
}

export interface TestLoginUsersResponse {
  items: TestLoginUser[];
  default_user_id: number;
}